Project structure modified: moved business logic from App to TodoContainer. Also reworked NavigationBar and AddTodoForm component
