+++ src/components/NavigationBar/NavigationBar.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 // Components
-import AddTodoForm from '../AddTodoForm/AddTodoForm';
 // Third party libraries:
 import PropTypes from 'prop-types';
 //CSS
 import styles from "./NavigationBar.module.css"
 
-const NavigationBar = ({onAddTodo, isLoading, inputIsFocused}) => {
+const NavigationBar = ({children}) => {
   return (
     <div className={styles.navBarContainer}>
       <nav className={styles.navBar}>
@@ -16,7 +15,7 @@ const NavigationBar = ({onAddTodo, isLoading, inputIsFocused}) => {
           </a>
         </div>
         <div className={styles.navBarUserControlsContainer}>
-          <AddTodoForm onAddTodo={onAddTodo} isLoading={isLoading} inputIsFocused={inputIsFocused}/>
+          {children}
         </div>
       </nav>
     </div>
@@ -33,8 +32,6 @@ export {
   ============================
 */
 NavigationBar.propTypes = {
-    onAddTodo: PropTypes.func.isRequired,
-    isLoading: PropTypes.bool.isRequired,
-    inputIsFocused: PropTypes.bool.isRequired
+  children: PropTypes.object.isRequired
 };
